// src/types/blog.ts
export interface BlogPost {
  slug: string;
  title: string;
  date: string;
  category: string;
  subCategory: string;
  tags: string[];
  readingTime: string;
  description: string;
  content: string;
}

export interface BlogMeta {
  slug: string;
  title: string;
  date: string;
  category: string;
  subCategory: string;
  tags: string[];
  readingTime: string;
  description: string;
}